+++ astropy/io/fits/tests/test_convenience.py
@@ -4,11 +4,10 @@ from __future__ import division, with_statement
 
 import warnings
 
-import numpy as np
-
 from ....extern import six  # pylint: disable=W0611
 from ....io import fits
 from ....table import Table
+from .. import printdiff
 from ....tests.helper import pytest, catch_warnings
 
 from . import FitsTestCase
@@ -52,7 +51,8 @@ class TestConvenience(FitsTestCase):
         with catch_warnings() as w:
             hdu = fits.table_to_hdu(table)
             assert len(w) == 1
-            assert str(w[0].message).startswith("'not-a-unit' did not parse as fits unit")
+            assert str(w[0].message).startswith("'not-a-unit' did not parse as"
+                                                " fits unit")
 
         # Check that TUNITn cards appear in the correct order
         # (https://github.com/astropy/astropy/pull/5720)
@@ -61,3 +61,36 @@ class TestConvenience(FitsTestCase):
         assert isinstance(hdu, fits.BinTableHDU)
         filename = str(tmpdir.join('test_table_to_hdu.fits'))
         hdu.writeto(filename, overwrite=True)
+
+    def test_printdiff(self):
+        """
+        Test that FITSDiff can run the different inputs without crashing.
+        """
+
+        # Testing different string input options
+        assert printdiff(self.data('arange.fits'),
+                         self.data('blank.fits')) is None
+        assert printdiff(self.data('arange.fits'),
+                         self.data('blank.fits'), ext=0) is None
+        assert printdiff(self.data('o4sp040b0_raw.fits'),
+                         self.data('o4sp040b0_raw.fits'),
+                         extname='sci') is None
+
+        # This may seem weird, but check printdiff to see, need to test
+        # incorrect second file
+        with pytest.raises(IOError):
+            printdiff('o4sp040b0_raw.fits', 'fakefile.fits', extname='sci')
+
+        # Test HDU object inputs
+        with fits.open(self.data('stddata.fits'), mode='readonly') as in1:
+            with fits.open(self.data('checksum.fits'), mode='readonly') as in2:
+
+                assert printdiff(in1[0], in2[0]) is None
+
+                with pytest.raises(ValueError):
+                    printdiff(in1[0], in2[0], ext=0)
+
+                assert printdiff(in1, in2) is None
+
+                with pytest.raises(NotImplementedError):
+                    printdiff(in1, in2, 0)